+++ modules/sheets/ageSystemSheetCharacter.js
@@ -539,8 +539,8 @@ export default class ageSystemSheetCharacter extends ActorSheet {
         let e = event.currentTarget;
         let itemId = e.dataset.itemId ?? e.closest(".feature-controls").dataset.itemId;
         const item = this.actor.items.get(itemId);
-        // return item.sheet.render(true); /** This line of code can return when I discover why Owned Items are not always opening the correct Item Sheet */
-        return item.openSheet();
+        return item.sheet.render(true); /** This line of code can return when I discover why Owned Items are not always opening the correct Item Sheet */
+        // return item.openSheet('age-system.ageSystemSheetItem');
     };
 
     _onItemDelete(event) {